Move CommUtil to protocol folder
