+++ casper/src/main/scala/coop/rchain/casper/blocks/proposer/Proposer.scala
@@ -10,9 +10,8 @@ import coop.rchain.blockstorage.dag.BlockDagStorage
 import coop.rchain.blockstorage.deploy.DeployStorage
 import coop.rchain.casper._
 import coop.rchain.casper.engine.BlockRetriever
-import coop.rchain.casper.protocol.BlockMessage
+import coop.rchain.casper.protocol.{BlockMessage, CommUtil}
 import coop.rchain.casper.syntax._
-import coop.rchain.casper.util.comm.CommUtil
 import coop.rchain.casper.util.rholang.RuntimeManager
 import coop.rchain.crypto.PrivateKey
 import coop.rchain.metrics.Metrics.Source
